# perf: make plugin load faster