### What I did

before:

```
CPU times: user 2.13 s, sys: 288 ms, total: 2.41 s
Wall time: 2.91 s
```

after:

```
CPU times: user 8.67 ms, sys: 4.3 ms, total: 13 ms
Wall time: 17.7 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)